types and directory restructure
