--- source/lang/src/frontend/token/freeArray.c
@@ -1,16 +1,0 @@
-#include <stdlib.h>
-#include "array.h"
-
-int freeArray( Array* array ) {
-    while ( ((Node*)array->head)->next->next != NULL ) {
-        deleteHead(array);
-        continue;
-    }
-    free(array->head);
-    array->head = NULL;
-    free(array->tail);
-    array->tail = NULL;
-    free(array);
-    array = NULL;
-    return 0;
-}
