+++ source/lang/src/frontend/types.h
@@ -1,0 +1,12 @@
+#ifndef _types_h_
+#define _types_h_
+
+/*#
+ * @brief Include directive for all the different types used by the iterpreter
+ *
+ */
+#include "typedef/array_t.h"
+#include "typedef/token_t.h"
+#include "typedef/astnode_t.h"
+
+#endif
